Assign ACE-27

Assign Adult Comorbidity Evaluation-27 index based on 26 medical comorbidities.

assign_ace27(
  ace_mi,
  ace_cad,
  ace_chf,
  ace_arr,
  ace_htn,
  ace_vd,
  ace_pad,
  ace_res,
  ace_hep,
  ace_sto,
  ace_pan,
  ace_rd,
  ace_dm,
  ace_str,
  ace_dem,
  ace_par,
  ace_neu,
  ace_psy,
  ace_rhe,
  ace_aid,
  ace_st,
  ace_lm,
  ace_lym,
  ace_alc,
  ace_id,
  ace_obe
)

Arguments

ace_mi
Myocardial Infarction

ace_cad
Angina / Coronary Heart Disease

ace_chf
Congestive Heart Failure

ace_arr
Arrhythmias

ace_htn
Hypertension

ace_vd
Venous Disease

ace_pad
Peripheral Arterial Disease

ace_res
Respiratory System

ace_hep
Hepatic

ace_sto
Stomach / Intestine

ace_pan
Pancreas

ace_rd
End-stage renal disease

ace_dm
Diabetes Mellitus

ace_str
Stroke

ace_dem
Dementia

ace_par
Paralysis

ace_neu
Neuromuscular

ace_psy
Psychiatric

ace_rhe
Rheumatologic

ace_aid
AIDS

ace_st
Solid Tumor including melanoma

ace_lm
Leukemia and Myeloma

ace_lym
Lymphoma

ace_alc
Alcohol

ace_id
Illicit Drugs

ace_obe
Obesity

Value

A factor with levels "None", "Mild", "Moderate", "Severe" indicating the overall ACE-27 score.

Details

Each comorbidity is scored on a scale of Grades 0-3, where 0 = None, 1 = Mild, 2 = Moderate, and 3 = Severe. The ACE-27 overall score is assigned according to the highest ranked individual comorbidity recorded. However, if there are two or more Grade 2 comorbidities occuring in different organ systems, then the overall score becomes Grade 3.

Two other scoring assumptions are made. First, if there is insufficient information regarding a comorbidity then it is described as an "NOS" (not otherwise specified) case, and scored as Grade 1. Secondly, any case with multiple symptoms from different Grades is scored into the higher Grade. For example, someone with symptoms from both Grade 1 and Grade 2 is categorized as Grade 2.

ACE comorbidities by organ systems

Cardiovascular System

Myocardial Infarction

Angina / Coronary Heart Disease

Congestive Heart Failure

Arrhythmias

Hypertension

Venous Disease

Peripheral Arterial Disease

Respiratory System

Gastrointestinal System

Hepatic

Stomach / Intestine

Pancreas

Renal System

End-stage renal disease

Endocrine System

Diabetes Mellitus

Neurological System

Stroke

Dementia

Paralysis

Neuromuscular

Psychiatric

Rheumatologic

Immunological System

AIDS

Malignancy

Solid Tumor including melanoma

Leukemia and Myeloma

Lymphoma

Substance Abuse

Alcohol

Illicit Drugs

Body Weight

Obesity

References

Binder PS, Peipert JF, Kallogjeri D, et al. Adult Comorbidity Evaluation 27 score as a predictor of survival in endometrial cancer patients. Am J Obstet Gynecol. 2016;215(6):766.e1-766.e9. doi:10.1016/j.ajog.2016.07.035

https://siog.org/files/public/ace27.pdf

Author

Derek Chiu