Assign ESMO 2013

Assign ESMO 2013 based on stage, grade, and histological subtype group

assign_esmo2013(stage_full, grade, hist_gr)

Arguments

stage_full

FIGO stage: I, II, III, IV with substages

grade

tumour grade: 1, 2, 3

hist_gr

histological subtype group: endometrioid or non-endometrioid

Value

ESMO 2013 assigned into "low", "intermediate", or "high". Unassignable cases are NA.

Details

ESMO 2013 is assigned using stage, grade, and histological subtype group into low, intermediate, and high risk based on the following criteria:

low:

stage IA, grade 1/2, with endometrioid type

intermediate:

stage IA, grade 3, with endometrioid type

stage IB, grade 1/2, with endometrioid type

high:

all stages with non-endometrioid type

stage IB, grade 3, with endometrioid type

stage II or higher

Note

Assignment starts from the high group first as the criteria are not mutually exclusive.

References

Colombo et al. Ann Oncol 2013

Author

Derek Chiu, Samuel Leung

Examples

esmo2013 <- with(emdb, assign_esmo2013(stage_full, grade_rev, hist_rev_gr))
table(esmo2013)
#> esmo2013
#>          low intermediate         high 
#>           89          117          425