Simulated endometrial cancer Database

Simulated endometrial cancer database used for testing functions.

emdb

Format

An object of class tbl_df (inherits from tbl, data.frame) with 800 rows and 16 columns.

Details

Each clinical variable has a distribution simulated from the proportions seen in the Vancouver endometrial cancer database.

stage_full

full FIGO staging

stage

stage without substages: I, II, III, IV

grade_rev

tumour grade

hist_rev_gr

histological subtype group

hist

histological subtype

myo

myometrial invasion

lvi

lymphovascular invasion (LVSI)

mmr_ihc_2

MMR status determined by 2 IHC markers

mmr_ihc_4

MMR status determined by 4 IHC markers

pole_mut

POLE mutation status

p53

p53 IHC

p53_mut

p53 mutation status

msh6, pms2, mlh1, msh2

MMR IHC markers

Examples

emdb

#> # A tibble: 800 × 16
#>    stage_full stage grade_rev hist_rev_gr  hist  myo   lvi   mmr_ihc_2 mmr_ihc_4
#>    <fct>      <fct> <fct>     <fct>        <fct> <fct> <fct> <fct>     <fct>    
#>  1 IA         I     grade 3   non-endomet… endo… none  nega… intact    intact   
#>  2 III        III   grade 3   endometrioid endo… 1-50% nega… intact    intact   
#>  3 IV         IV    grade 1   endometrioid sero… >50%  posi… intact    intact   
#>  4 IA         I     grade 3   endometrioid sero… 1-50% nega… intact    intact   
#>  5 II         II    grade 1   non-endomet… clea… none  nega… intact    intact   
#>  6 IV         IV    grade 1   endometrioid clea… 1-50% nega… intact    intact   
#>  7 II         II    grade 3   non-endomet… endo… 1-50% nega… deficient deficient
#>  8 IA         I     grade 3   endometrioid sero… 1-50% nega… intact    intact   
#>  9 III        III   grade 1   non-endomet… endo… 1-50% posi… intact    intact   
#> 10 IVB        IV    grade 3   non-endomet… endo… >50%  nega… intact    intact   
#> # … with 790 more rows, and 7 more variables: pole_mut <fct>, p53 <fct>,
#> #   p53_mut <fct>, msh6 <fct>, pms2 <fct>, mlh1 <fct>, msh2 <fct>