Assign ESMO 2016

Assign ESMO 2016 based on stage, grade, histological subtype group, myometrial invasion, and LVSI. Residual disease can be used if available.

assign_esmo2016(stage_full, grade, hist_gr, myo, lvi, residual = NULL)

Arguments

stage_full
FIGO stage: I, II, III, IV with substages

grade
tumour grade: 1, 2, 3

hist_gr
histological subtype group: endometrioid or non-endometrioid

myo
myometrial invasion: none, 1-50%, >50%

lvi
lymphovascular space invasion: negative, positive, focal, extensive

residual
residual disease: "<1cm", ">=1cm", "microscopic", or "no residual". Use if available. Microscopic residual is regarded as no residual disease.

Value

ESMO 2016 assigned into "low", "intermediate", "high-intermediate", "high", "advanced", or "metastatic". Unassignable cases are NA.

Details

ESMO 2016 is assigned using stage, grade, histological subtype group, myometrial invasion, and LVSI into low, intermediate, high-intermediate, high, advanced, and metastatic risk based on the following criteria:

low:

stage I, grade 1/2, endometrioid, <50% myometrial invasion, LVSI negative

intermediate:

stage I, grade 1/2, endometrioid, >=50% myometrial invasion, LVSI negative

high-intermediate:

stage I, grade 3, endometrioid, <50% myometrial invasion, regardless of LVSI status

stage I, grade 1/2, endometrioid, LVSI unequivocally positive, regardless of depth of invasion

high:

stage I, grade 3, endometrioid, >=50% myometrial invasion, regardless of LVSI status

stage II or higher

stage III, endometrioid, (no residual disease if known)

non-endometrioid (serous or clear-cell or undifferentiated carcinoma, or carcinosarcoma), (no residual disease if known)

advanced:

stage III, (with residual disease if known)

stage IVA

metastatic:

stage IVB

Note

Assignment starts from the metastatic group first as the criteria are not mutually exclusive.

References

Colombo et al. Annals of Oncology 2016

Author

Derek Chiu, Samuel Leung

Examples

esmo2016 <- with(emdb, assign_esmo2016(stage_full, grade_rev, hist_rev_gr,
myo, lvi, residual))
table(esmo2016)

#> esmo2016
#>               low      intermediate high-intermediate              high 
#>                88                56               174               448 
#>          advanced        metastatic 
#>                28                 6