Assign ESMO 2020

Assign ESMO 2020 based on stage, grade, histological subtype group, myometrial invasion, and LVSI. Molecular classification and residual disease can be used if available.

assign_esmo2020(
  stage_full,
  grade,
  hist_gr,
  myo,
  lvi,
  eclass = NULL,
  residual = NULL
)

Arguments

stage_full
FIGO stage: I, II, III, IV with substages

grade
tumour grade: 1, 2, 3

hist_gr
histological subtype group: endometrioid or non-endometrioid

myo
myometrial invasion: none, 1-50%, >50%

lvi
lymphovascular space invasion: negative, positive, focal, extensive

eclass
molecular classification: "MMRd", "POLEmut", "p53abn", or "NSMP/p53wt"

residual
residual disease: "<1cm", ">=1cm", "microscopic", or "no residual". Use if available. Microscopic residual is regarded as no residual disease.

Value

ESMO 2020 assigned into "low", "intermediate", "high-intermediate", "high", "advanced", or "metastatic". Unassignable cases are NA.

Details

ESMO 2020 is assigned using stage, grade, histological subtype group, myometrial invasion, and LVSI into low, intermediate, high-intermediate, high, advanced, and metastatic risk based on the following criteria with and without molecular classification/residual disease:

With molecular classification

low:

stage I-II, POLEmut, (no residual disease if known)

missing residual disease is regarded as no residual disease

stage IA, grade 1/2, endometrioid, LVSI negative or focal, MMRd/NSMP

intermediate:

stage IB, grade 1/2, endometrioid, LVSI negative or focal, MMRd/NSMP

stage IA, grade 3, endometrioid, LVSI negative or focal, MMRd/NSMP

stage IA, no myometrial invasion, p53abn

high-intermediate:

stage IA/IB, endometrioid, LVSI positive or extensive, MMRd/NSMP

stage IB, grade 3, endometrioid, MMRd/NSMP

stage II/IIA, endometrioid, MMRd/NSMP

high:

stage III-IVA, endometrioid, MMRd/NSMP, (no residual disease if known)

stage I-IVA, >0% myometrial invasion, p53abn, (no residual disease if known)

for stage I-II, missing residual disease is regarded as no residual disease

stage IA, non-endometrioid or mixed subtype, >0% myometrial invasion, MMRd/NSMP, (no residual disease if known)

missing residual disease is regarded as no residual disease

stage IB-IVA, non-endometrioid or mixed subtype, MMRd/NSMP, (no residual disease if known)

for stage IB-II, missing residual disease is regarded as no residual disease

advanced:

stage III-IVA, any molecular type, (with residual disease if known)

metastatic:

stage IVB, any molecular type

Without molecular classification

low:

stage IA, grade 1/2, endometrioid, LVSI negative or focal

intermediate:

stage IB, grade 1/2, endometrioid, LVSI negative or focal

stage IA, grade 3, endometrioid, LVSI negative or focal

stage IA, no myometrial invasion

high-intermediate:

stage IA/IB, endometrioid, LVSI positive or extensive

stage IB, grade 3, endometrioid

stage II/IIA, endometrioid

high:

stage III-IVA, endometrioid, (no residual disease if known)

stage IA, non-endometrioid or mixed subtype, >0% myometrial invasion, (no residual disease if known)

missing residual disease is regarded as no residual disease

stage IB-IVA, non-endometrioid or mixed subtype, (no residual disease if known)

for stage IB-II, missing residual disease is regarded as no residual disease

advanced:

stage III-IVA, (with residual disease if known)

metastatic:

stage IVB

Note

Assignment starts from the low group first.

References

Concin N, Matias-Guiu X, Vergote I, et al ESGO/ESTRO/ESP guidelines for the management of patients with endometrial carcinoma. International Journal of Gynecologic Cancer Published Online First: 18 December 2020. doi: 10.1136/ijgc-2020-002230

Author

Derek Chiu

Examples

# without molecular classification
esmo2020_wo_eclass <- with(emdb, assign_esmo2020(stage_full, grade_rev,
hist_rev_gr, myo, lvi))
table(esmo2020_wo_eclass)
#> esmo2020_wo_eclass
#>               low      intermediate high-intermediate              high 
#>                51               100               130               285 
#>          advanced        metastatic 
#>                17                 6 

# with molecular classification
eclass <- with(emdb, assign_promise2015(mmr_ihc_2, pole_mut, p53))
esmo2020_w_eclass <- with(emdb, assign_esmo2020(stage_full, grade_rev,
hist_rev_gr, myo, lvi, eclass))
table(esmo2020_w_eclass)
#> esmo2020_w_eclass
#>               low      intermediate high-intermediate              high 
#>                75                53                89               265 
#>          advanced        metastatic 
#>                38                 6 

# with molecular classification and residual disease
esmo2020_w_eclass_res <- with(emdb, assign_esmo2020(stage_full, grade_rev,
hist_rev_gr, myo, lvi, eclass, residual))
table(esmo2020_w_eclass_res)
#> esmo2020_w_eclass_res
#>               low      intermediate high-intermediate              high 
#>                70                53                89               239 
#>          advanced        metastatic 
#>                32                 6