Simulated endometrial cancer Database

Simulated endometrial cancer database used for testing functions.

emdb

Format

An object of class tbl_df (inherits from tbl, data.frame) with 800 rows and 43 columns.

Details

Each clinical variable has a distribution simulated from the proportions seen in the Vancouver endometrial cancer database.

stage_full

full FIGO staging

stage

stage without substages: I, II, III, IV

grade_rev

tumour grade

hist_rev_gr

histological subtype group

hist

histological subtype

myo

myometrial invasion

lvi

lymphovascular invasion (LVSI)

residual

residual disease

mmr_ihc_2

MMR status determined by 2 IHC markers

mmr_ihc_4

MMR status determined by 4 IHC markers

pole_mut

POLE mutation status

p53

p53 IHC

p53_mut

p53 mutation status

msh6, pms2, mlh1, msh2

MMR IHC markers

Examples

emdb
#> # A tibble: 800 × 43
#>    stage_full stage grade_rev hist_rev_gr  hist   myo   lvi   residual mmr_ihc_2
#>    <fct>      <fct> <fct>     <fct>        <fct>  <fct> <fct> <fct>    <fct>    
#>  1 IA         I     grade 3   non-endomet… endom… none  NA    no resi… intact   
#>  2 III        III   grade 3   NA           endom… 1-50% nega… no resi… intact   
#>  3 IV         IV    grade 1   endometrioid serous >50%  posi… no resi… intact   
#>  4 IA         NA    grade 3   endometrioid serous 1-50% nega… no resi… intact   
#>  5 II         II    grade 1   non-endomet… clear… none  nega… no resi… intact   
#>  6 IV         IV    grade 1   endometrioid clear… 1-50% nega… no resi… intact   
#>  7 II         II    grade 3   non-endomet… endom… 1-50% nega… NA       intact   
#>  8 IA         I     grade 3   endometrioid serous 1-50% nega… no resi… deficient
#>  9 III        III   grade 1   non-endomet… endom… 1-50% posi… no resi… intact   
#> 10 IVB        NA    grade 3   non-endomet… endom… >50%  nega… no resi… intact   
#> # … with 790 more rows, and 34 more variables: mmr_ihc_4 <fct>, pole_mut <fct>,
#> #   p53 <fct>, p53_mut <fct>, msh6 <fct>, pms2 <fct>, mlh1 <fct>, msh2 <fct>,
#> #   ace_mi <chr>, ace_cad <chr>, ace_chf <chr>, ace_arr <chr>, ace_htn <chr>,
#> #   ace_vd <chr>, ace_pad <chr>, ace_res <chr>, ace_hep <chr>, ace_sto <chr>,
#> #   ace_pan <chr>, ace_rd <chr>, ace_dm <chr>, ace_str <chr>, ace_dem <chr>,
#> #   ace_par <chr>, ace_neu <chr>, ace_psy <chr>, ace_rhe <chr>, ace_aid <chr>,
#> #   ace_st <chr>, ace_lm <chr>, ace_lym <chr>, ace_alc <chr>, ace_id <chr>, …